+++ rust/src/wireguard_runtime/tcp.rs
@@ -6,6 +6,7 @@ use bytes::Bytes;
 use smoltcp::iface::{Config, Interface, SocketHandle, SocketSet};
 use smoltcp::socket::tcp;
 use smoltcp::wire::{HardwareAddress, IpAddress};
+use tokio::runtime::Handle;
 use tokio::sync::mpsc;
 use tracing::{debug, error};
 
@@ -39,11 +40,12 @@ pub struct TcpInterface {
 
 impl TcpInterface {
     pub fn spawn(
+        handle: &Handle,
         source_ips: Vec<IpAddr>,
         device: ChannelIpDevice,
         commands: mpsc::UnboundedReceiver<TcpCommand>,
     ) {
-        tokio::spawn(async move {
+        handle.spawn(async move {
             let runtime = Self {
                 source_ips,
                 device,